requirements.txt
anthropic>=0.40.0
tavily-python>=0.5.0
playwright>=1.40.0
jinja2>=3.1.0
markdown>=3.5.0
slack-sdk>=3.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
